pyproject.toml
[project]
name = "asistente-gastos"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.2.0",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.14.8",
    "pytest>=8.4.2",
]
